Changed the data type for shares from Int to Decimal, because my mutual funds in the RRSP account are able to have fractional share counts and maybe I will use mutual funds again some day.
Updated the database tables as well and manually edited the EF model.

+++ StockPortfolioApplication/Equity.cs
@@ -13,7 +13,7 @@ namespace StockPortfolioApplication
         public int ID { get; set; }
         public string Ticker { get; set; }
         public string Description { get; set; }
-        public int Shares { get; set; }
+        public decimal Shares { get; set; }
         public decimal CurrentPrice { get; set; }
 
         public decimal AverageCostBasis { get; set; }
